requirements.txt
discord.py==2.0.1
pillow==9.5.0
mss==7.0.1
pytesseract==0.3.9
